Moves Settings from Editor to Settings
